requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.13
boto3>=1.33.0
botocore>=1.33.0
qdrant-client==1.7.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==9.3.0
PyMuPDF==1.21.1
python-multipart==0.0.5
pydantic==1.10.12 